deploy/src/client/requirements.txt
Pillow>=10.4,<12.0
